+++ components/common/PageLayout/components/NewPageMenu.tsx
@@ -1,7 +1,6 @@
 import styled from '@emotion/styled';
 import AddIcon from '@mui/icons-material/Add';
 import ArticleIcon from '@mui/icons-material/InsertDriveFileOutlined';
-import DatabaseIcon from '@mui/icons-material/TableChart';
 import IconButton from '@mui/material/IconButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import Menu from '@mui/material/Menu';
@@ -11,6 +10,7 @@ import Typography from '@mui/material/Typography';
 import { Page } from 'models';
 import { MouseEvent, useState } from 'react';
 import { greyColor2 } from 'theme/colors';
+import { StyledDatabaseIcon } from './PageIcon';
 
 export const StyledIconButton = styled(IconButton)`
   border-radius: 3px;
@@ -26,12 +26,6 @@ export const StyledIconButton = styled(IconButton)`
 `;
 
 export const StyledArticleIcon = styled(ArticleIcon)`
-  color: ${greyColor2};
-  opacity: 0.5;
-  font-size: 22px;
-`;
-
-export const StyledDatabaseIcon = styled(DatabaseIcon)`
   color: ${greyColor2};
   opacity: 0.5;
   font-size: 22px;
